pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "csv-cleaner"
version = "0.1.0"
description = "Config-driven, modular CSV cleaning engine"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
csv-cleaner = "cleaner.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cleaner*"]

[tool.ruff]
line-length = 100
target-version = "py39"
